+++ src/test/java/wolox/training/BookControllerTest.java
@@ -11,27 +11,23 @@ import org.springframework.http.MediaType;
 import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
-import wolox.training.controllers.UsersController;
+import wolox.training.controllers.BookController;
 import wolox.training.models.Book;
-import wolox.training.models.Users;
-import wolox.training.repositories.UsersRepository;
-
-import java.time.LocalDate;
+import wolox.training.repositories.BookRepository;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @RunWith(SpringRunner.class)
-@WebMvcTest(UsersController.class)
-public class UserRestControllerIntegrationTest {
+@WebMvcTest(BookController.class)
+public class BookControllerTest {
     @Autowired
     private MockMvc mvc;
 
     @MockBean
-    private UsersRepository mockUserRepository;
+    private BookRepository mockBookRepository;
 
-    private Users oneTestUser;
     private Book oneTestBook;
 
     @Before
@@ -46,30 +42,33 @@ public class UserRestControllerIntegrationTest {
                 "Bantam books",
                 1998
         );
-        oneTestUser = new Users(
-                "ocolmenares",
-                "Oriana",
-                "password",
-                LocalDate.parse("1997-11-03")
-        );
-        oneTestUser.addBook(oneTestBook);
     }
 
     @WithMockUser
     @Test
-    public void whenFindByIdWhichExist_thenUserIsReturned() throws Exception {
-        Mockito.when(mockUserRepository.findById(1L)).thenReturn(java.util.Optional.ofNullable(oneTestUser));
-        String url = ("/api/users/1");
+    public void whenFindByIdWhichExist_thenBookIsReturned() throws Exception {
+        Mockito.when(mockBookRepository.findById(1L)).thenReturn(java.util.Optional.ofNullable(oneTestBook));
+        String url = ("/api/books/1");
         mvc.perform(get(url)
                 .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andExpect(jsonPath
-                        ("$.username").value("ocolmenares"))
+                        ("$.title").value("Song of ice and fire"))
                 .andExpect(jsonPath
-                        ("$.name").value("Oriana"))
+                        ("$.isbn").value("1234"))
                 .andExpect(jsonPath
-                        ("$.password").value("password"))
+                        ("$.author").value("G.R.Martin"))
                 .andExpect(jsonPath
-                        ("$.birthday").value("1997-11-03"));
+                        ("$.genre").value("Fantasy"))
+                .andExpect(jsonPath
+                        ("$.subtitle").value("A Game of Thrones"))
+                .andExpect(jsonPath
+                        ("$.image").value("insert image"))
+                .andExpect(jsonPath
+                        ("$.publisher").value("Bantam books"))
+                .andExpect(jsonPath
+                        ("$.year").value(1998)
+                );
     }
 }
+
