+++ src/test/java/wolox/training/UserControllerTest.java
@@ -11,23 +11,27 @@ import org.springframework.http.MediaType;
 import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
-import wolox.training.controllers.BookController;
+import wolox.training.controllers.UsersController;
 import wolox.training.models.Book;
-import wolox.training.repositories.BookRepository;
+import wolox.training.models.Users;
+import wolox.training.repositories.UsersRepository;
+
+import java.time.LocalDate;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @RunWith(SpringRunner.class)
-@WebMvcTest(BookController.class)
-public class BookRestControllerIntegrationTest {
+@WebMvcTest(UsersController.class)
+public class UserControllerTest {
     @Autowired
     private MockMvc mvc;
 
     @MockBean
-    private BookRepository mockBookRepository;
+    private UsersRepository mockUserRepository;
 
+    private Users oneTestUser;
     private Book oneTestBook;
 
     @Before
@@ -42,33 +46,30 @@ public class BookRestControllerIntegrationTest {
                 "Bantam books",
                 1998
         );
+        oneTestUser = new Users(
+                "ocolmenares",
+                "Oriana",
+                "password",
+                LocalDate.parse("1997-11-03")
+        );
+        oneTestUser.addBook(oneTestBook);
     }
 
     @WithMockUser
     @Test
-    public void whenFindByIdWhichExist_thenBookIsReturned() throws Exception {
-        Mockito.when(mockBookRepository.findById(1L)).thenReturn(java.util.Optional.ofNullable(oneTestBook));
-        String url = ("/api/books/1");
+    public void whenFindByIdWhichExist_thenUserIsReturned() throws Exception {
+        Mockito.when(mockUserRepository.findById(1L)).thenReturn(java.util.Optional.ofNullable(oneTestUser));
+        String url = ("/api/users/1");
         mvc.perform(get(url)
                 .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andExpect(jsonPath
-                        ("$.title").value("Song of ice and fire"))
+                        ("$.username").value("ocolmenares"))
                 .andExpect(jsonPath
-                        ("$.isbn").value("1234"))
+                        ("$.name").value("Oriana"))
                 .andExpect(jsonPath
-                        ("$.author").value("G.R.Martin"))
+                        ("$.password").value("password"))
                 .andExpect(jsonPath
-                        ("$.genre").value("Fantasy"))
-                .andExpect(jsonPath
-                        ("$.subtitle").value("A Game of Thrones"))
-                .andExpect(jsonPath
-                        ("$.image").value("insert image"))
-                .andExpect(jsonPath
-                        ("$.publisher").value("Bantam books"))
-                .andExpect(jsonPath
-                        ("$.year").value(1998)
-                );
+                        ("$.birthday").value("1997-11-03"));
     }
 }
-
